Create Python_1.py
This is a Python code that assigns a grade to the user when they enter a value for a grade. If the value is greater than or equal to 90, it will give an "A" grade. If the value is greater than or equal to 80, it will give a "B" grade, and if the value is greater than or equal to 70, it will give a "C" grade.